requirements.txt
fastai==2.6.3
Flask==2.1.2
googlemaps==4.6.0
mysql_connector_repackaged==0.3.1
pandas==1.3.4
Pillow==9.1.1
requests==2.26.0
Werkzeug==2.1.2
